refactor: move DimensionLike into its own file
